Details
I followed a guide
(http://cwbuecheler.com/web/tutorials/2013/node-express-mongo/#) that
uses the express-generator. This works but is very messy as it includes
many things we will not use. We should clean this up or find an easier
way for next time. To run this, cd into nodetest1 in a command prompt
and type "npm start" to run the server. Then download mongo, cd into the
bin and type mongod --dbpath "C:\(Path to nodetest1/data)" to start the
datebase. The page should be on http://localhost:3000/newuser